requirements.txt
numpy==1.26.2
Pillow==10.1.0
streamlit==1.26.0
